fix: [UF-196] 배경 프로바이더 뷰포트에서만 나타나는 에러 수정 #175
